% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_meta.R
\name{collapse_authors}
\alias{collapse_authors}
\title{Collapse metadata table with author information to a single row per
author}
\usage{
collapse_authors(tib, sep_author = "; ", sep_address = "|")
}
\arguments{
\item{tib}{Data frame. Publication records, typically the
output of `id_to_meta`.}

\item{sep_author}{Character. The separator to use between different authors.}

\item{sep_address}{Character. The separator to use between different author
addresses.}
}
\value{
Data frame where each row corresponds to a different record.
}
\description{
Converts a data frame of publication records from a single row
per author to a single row per record. Author information is collapsed to a
single cell, with a given separator.
}
\examples{
\dontrun{ 
collapse_authors(tib, sep_author = " and ", sep_address = "|")
}
}
\author{
Vladislav Nachev \email{vladislav.nachev@charite.de}
}
